implement PAP and got to sync events for all protocols
sync event are nesseary because the result of processing
a frame has to be applied to the main fsm before the next
packet can be processed (e.g. we need to go to the auth
pahse before we can process PAP frames). With async events
that ordering could no be guaranteed.

Certain protocol events (mainly timeouts) still need to be
delivered asynchronously.